+++ core/src/main/java/org/elasticsearch/index/query/CommonTermsQueryParser.java
@@ -21,7 +21,6 @@ package org.elasticsearch.index.query;
 
 import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.xcontent.XContentParser;
-import org.elasticsearch.index.query.QueryBuilder.Operator;
 
 import java.io.IOException;
